requirements.txt
Django>=5.0
psycopg2-binary
django-allauth
django-storages[boto3]
django-ratelimit
django-csp
django-environ
django-watson
celery[redis]
redis
python-magic
bleach
Pillow
factory-boy
pytest-django
black
isort
flake8
